+++ src/mako/http/routing/attributes/Middleware.php
@@ -15,19 +15,43 @@ use Attribute;
 #[Attribute(Attribute::IS_REPEATABLE | Attribute::TARGET_CLASS | Attribute::TARGET_METHOD)]
 class Middleware
 {
+	/**
+	 * Parameters.
+	 */
+	protected array $parameters;
+
     /**
      * Constructor.
      */
     public function __construct(
-        protected array|string $middleware
+        protected string $middleware,
+		mixed ...$parameters
     )
-    {}
+    {
+		$this->parameters = $parameters;
+	}
 
     /**
-     * Returns an array of middleware.
+     * Returns the middleware.
      */
-    public function getMiddleware(): array
+    public function getMiddleware(): string
     {
-        return (array) $this->middleware;
+        return $this->middleware;
     }
+
+	/**
+	 * Returns the middleware parameters.
+	 */
+	public function getParameters(): array
+	{
+		return $this->parameters;
+	}
+
+	/**
+	 * Returns the middleware and parameters.
+	 */
+	public function getMiddlewareAndParameters(): array
+	{
+		return ['middleware' => $this->middleware, 'parameters' => $this->parameters];
+	}
 }