+++ tests/unit/http/routing/attributes/ConstraintTest.php
@@ -5,7 +5,7 @@
  * @license   http://www.makoframework.com/license
  */
 
-namespace mako\tests\unit\http\routing\middleware;
+namespace mako\tests\unit\http\routing\attributes;
 
 use mako\http\routing\attributes\Constraint;
 use mako\tests\TestCase;
@@ -18,20 +18,30 @@ class ConstraintTest extends TestCase
 	/**
 	 *
 	 */
-	public function testWithArray(): void
+	public function testGetConstraint(): void
 	{
-		$constraint = new Constraint(['foobar', 'barfoo']);
+		$constraint = new Constraint(FooBar::class);
 
-		$this->assertSame(['foobar', 'barfoo'], $constraint->getConstraints());
+		$this->assertSame(FooBar::class, $constraint->getConstraint());
 	}
 
 	/**
 	 *
 	 */
-	public function testWithString(): void
+	public function testGetParameters(): void
 	{
-		$constraint = new Constraint('foobar');
+		$constraint = new Constraint(FooBar::class, foo: 'bar');
 
-		$this->assertSame(['foobar'], $constraint->getConstraints());
+		$this->assertSame(['foo' => 'bar'], $constraint->getParameters());
+	}
+
+	/**
+	 *
+	 */
+	public function testGetConstraintAndParameters(): void
+	{
+		$constraint = new Constraint(FooBar::class, foo: 'bar');
+
+		$this->assertSame(['constraint' => FooBar::class, 'parameters' => ['foo' => 'bar']], $constraint->getConstraintAndParameters());
 	}
 }