+++ CHANGELOG.md
@@ -22,6 +22,8 @@ The major version bump is due to dropped support for PHP `8.0` and a several bre
 	- `JSON::setStatus()`
 	- `Redirect::setStatus()`
 * Added `HttpStatusException::getStatus()` method.
+* Improved middleware and constraint registration.
+* Global middleware and constraints can now be registered with parameters.
 
 #### Changes